Move inspect writer methods to Inspect iterator
Remove cargo feature-gated `Symbol::format_inspect_into` and
`Symbol::write_inspect_into` methods from `Symbol` and add
`Inspect::format_into` and `Inspect::write_into`.

This commit makes it possible to write examples and doctests for these
APIs.
