+++ spinoso-symbol/src/lib.rs
@@ -324,79 +324,4 @@ impl Symbol {
             Inspect::default()
         }
     }
-
-    /// Write an [`Inspect`] iterator into the given destination using the debug
-    /// representation of the interned byteslice associated with the symbol in
-    /// the underlying interner.
-    ///
-    /// This formatter writes content like `:spinoso` and `:"invalid-\xFF-utf8"`.
-    /// To see example output of the underlying iterator, see the [`Inspect`]
-    /// documentation.
-    ///
-    /// To write binary output, use [`write_inspect_into`], which requires the
-    /// `std` feature to be activated.
-    ///
-    /// If there symbol does not exist in the underlying interner or there is an
-    /// error looking up the symbol in the underlying interner, a default
-    /// inspect representation is written.
-    ///
-    /// # Errors
-    ///
-    /// If the given writer returns an error as it is being written to, that
-    /// error is returned.
-    ///
-    /// [`write_inspect_into`]: Self::write_inspect_into
-    #[inline]
-    #[cfg(feature = "artichoke")]
-    #[cfg_attr(docsrs, doc(cfg(feature = "artichoke")))]
-    pub fn format_inspect_into<T, U, W>(self, interner: &T, mut dest: W) -> fmt::Result
-    where
-        T: Intern<Symbol = U>,
-        U: Copy + From<Symbol>,
-        W: fmt::Write,
-    {
-        let inspect = self.inspect(interner);
-        for ch in inspect {
-            dest.write_char(ch)?;
-        }
-        Ok(())
-    }
-
-    /// Write an [`Inspect`] iterator into the given destination using the debug
-    /// representation of the interned byteslice associated with the symbol in
-    /// the underlying interner.
-    ///
-    /// This formatter writes content like `:spinoso` and `:"invalid-\xFF-utf8"`.
-    /// To see example output of the underlying iterator, see the [`Inspect`]
-    /// documentation.
-    ///
-    /// To write to a `String` formatter, use [`format_inspect_into`].
-    ///
-    /// If there symbol does not exist in the underlying interner or there is an
-    /// error looking up the symbol in the underlying interner, a default
-    /// inspect representation is written.
-    ///
-    /// # Errors
-    ///
-    /// If the given writer returns an error as it is being written to, that
-    /// error is returned.
-    ///
-    /// [`format_inspect_into`]: Self::format_inspect_into
-    #[inline]
-    #[cfg(all(feature = "artichoke", feature = "std"))]
-    #[cfg_attr(docsrs, doc(cfg(all(feature = "artichoke", feature = "std"))))]
-    pub fn write_inspect_into<T, U, W>(self, interner: &T, mut dest: W) -> std::io::Result<()>
-    where
-        T: Intern<Symbol = U>,
-        U: Copy + From<Symbol>,
-        W: std::io::Write,
-    {
-        let inspect = self.inspect(interner);
-        let mut buf = [0; 4];
-        for ch in inspect {
-            let utf8 = ch.encode_utf8(&mut buf);
-            dest.write_all(utf8.as_bytes())?;
-        }
-        Ok(())
-    }
 }
